fix: align RITS tests with CI checks
