+++ src/exgentic/agents/litellm_tool_calling/instance.py
@@ -140,8 +140,7 @@ class LiteLLMToolCallingAgentInstance(AgentInstance):
             self._cost_data.update_cost_from_tokens(usage.prompt_tokens, usage.completion_tokens)
         except Exception:
             self.logger.debug(
-                "Cost tracking not available for model %s "
-                "(prompt_tokens=%s completion_tokens=%s)",
+                "Cost tracking not available for model %s (prompt_tokens=%s completion_tokens=%s)",
                 self.model,
                 usage.prompt_tokens,
                 usage.completion_tokens,
